% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raincloud.R
\name{geom_flat_violin}
\alias{geom_flat_violin}
\title{Flat Violin Plot}
\usage{
geom_flat_violin(
  mapping = NULL,
  data = NULL,
  stat = "ydensity",
  position = "dodge",
  trim = TRUE,
  scale = "area",
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by aes() or aes_().}

\item{data}{The data to be displayed in this layer.}

\item{stat}{Which statistic. Defaults to "ydensity"}

\item{position}{Position adjustment, either as a string, or the result of a call to a position adjustment function.}

\item{trim}{If FALSE, the default, each density is computed on the full range of the data. If TRUE, each density is computed over the range of that group: this typically means the estimated x values will not line-up, and hence you won't be able to stack density values. This parameter only matters if you are displaying multiple densities in one plot or if you are manually adjusting the scale limits.}

\item{scale}{if "area" (default), all violins have the same area (before trimming the tails). If "count", areas are scaled proportionally to the number of observations. If "width", all violins have the same maximum width.}

\item{show.legend}{Whether or not to show the legend.}

\item{inherit.aes}{If mapping is specified and inherit.aes = TRUE (the default), the mapping is combined with the default mapping at the top level of the plot.}

\item{...}{Other arguments passed on to layer(). These are often aesthetics, used to set an aesthetic to a fixed value, like colour = "red" or size = 3. They may also be parameters to the paired geom/stat.}
}
\description{
Flat Violin Plot
}
\examples{
ggplot(iris, aes(Species, Sepal.Width)) + geom_flat_violin()
}
