% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitviolin.R
\name{geom_split_violin}
\alias{geom_split_violin}
\title{Split Violin Plot}
\usage{
geom_split_violin(
  mapping = NULL,
  data = NULL,
  stat = "ydensity",
  position = "identity",
  ...,
  draw_quantiles = NULL,
  trim = TRUE,
  scale = "area",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by aes() or aes_().}

\item{data}{The data to be displayed in this layer.}

\item{stat}{Which statistic. Defaults to "ydensity"}

\item{position}{Position adjustment, either as a string, or the result of a call to a position adjustment function.}

\item{...}{Other arguments passed on to layer(). These are often aesthetics, used to set an aesthetic to a fixed value, like colour = "red" or size = 3. They may also be parameters to the paired geom/stat.}

\item{draw_quantiles}{If not(NULL) (default), draw horizontal lines at the given quantiles of the density estimate.}

\item{trim}{If TRUE (default), trim the tails of the violins to the range of the data. If FALSE, don't trim the tails.}

\item{scale}{if "area" (default), all violins have the same area (before trimming the tails). If "count", areas are scaled proportionally to the number of observations. If "width", all violins have the same maximum width.}

\item{na.rm}{If FALSE, the default, missing values are removed with a warning. If TRUE, missing values are silently removed.}

\item{show.legend}{Whether or not to show the legend.}

\item{inherit.aes}{If mapping is specified and inherit.aes = TRUE (the default), the mapping is combined with the default mapping at the top level of the plot.}
}
\description{
Split Violin Plot
}
\examples{
ggplot(ggplot2::diamonds, aes(cut, carat)) + geom_split_violin()
}
