% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raincloud.R
\name{\%||\%}
\alias{\%||\%}
\title{If Null}
\usage{
a \%||\% b
}
\arguments{
\item{a}{value to check if NULL}

\item{b}{replacement if NULL}
}
\value{
value or replacement
}
\description{
If Null
}
\keyword{internal}
