% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldt_data.R
\docType{data}
\name{ldt_data}
\alias{ldt_data}
\title{Stroop Task Example Data}
\format{
A data frame with 100 rows and 7 variables:
\describe{
\item{id}{Participant ID}
\item{age}{Age of participant in years}
\item{language}{Languages (1 = monolingual, 2 = bilingual)}
\item{rt_word}{Mean reaction time for words}
\item{rt_nonword}{Mean reaction time for non-words}
\item{acc_word}{Accuracy for words}
\item{acc_nonword}{Accuracy for non-words}
}
}
\source{
\url{}
}
\usage{
ldt_data
}
\description{
Simulated stroop task data for monolingual and bilingual participants. DVs are mean reaction time (rt) and accuracy (acc) for words and non-words.
}
\keyword{datasets}
