% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldt_long.R
\docType{data}
\name{ldt_long}
\alias{ldt_long}
\title{Stroop Task Example Data - long format}
\format{
A data frame with 200 rows and 6 variables:
\describe{
\item{id}{Participant ID}
\item{age}{Age of participant in years}
\item{language}{Languages (monolingual, bilingual)}
\item{condition}{Trial condition (word, nonword)}
\item{rt}{Mean reaction time (ms)}
\item{acc}{Accuracy}
}
}
\usage{
ldt_long
}
\description{
Simulated stroop task data for monolingual and bilingual participants. DVs are mean reaction time (rt) and accuracy (acc) for words and non-words.
}
\keyword{datasets}
